requirements.txt
# Vector + graph memory
lancedb>=0.6.0
kuzu>=0.4.0

# Embeddings (local, no API)
sentence-transformers>=3.0.0

# Prompt compression
llmlingua>=0.2.0

# Observability
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-otlp>=1.24.0

# Evals
deepeval>=1.0.0

# HTTP + utils
httpx>=0.27.0
pydantic>=2.0.0
rich>=13.0.0
click>=8.1.0
python-dotenv>=1.0.0

# System monitoring
psutil>=5.9.0

# Google Drive integration (cold storage)
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.90.0

# Telegram bot (remote communication)
python-telegram-bot>=21.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
